Sorry,forget to apply some fixed. and [1612] has fixed speed changes.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@1614 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/shared/Database/DBCStructure.h
@@ -103,7 +103,7 @@ struct SpellEntry
     uint32    EffectRadiusIndex[3];
     uint32    EffectApplyAuraName[3];
     uint32    EffectAmplitude[3];
-    uint32    EffectMultipleValue[3];
+    float     EffectMultipleValue[3];
     uint32    EffectChainTarget[3];
     uint32    EffectItemType[3];
     uint32    EffectMiscValue[3];
